Object ID's fixed for Svija Life
Previously all object ID's, which normally begin SVGID_, were renamed to match the parent SVG.

This was done to avoid a situation where multiple SVG's on a page would cause ID conflicts.

But, it prevented Svija Life from being able to easily sort named and unnammed objects in SVG files.

This fix adds the custom part of the ID *after* the SVGID_.